+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Split out from [pureharm](https://github.com/busymachines/pureharm) as of versio
 
 Newly cross published for both Scala 2.13, and 3.0.0-RC1 on JVM and JS runtimes.
 
+- add instance for Show[Throwable] to `busymachines.pureharm.PureharmCoreImplicits`
+
 `busymachines.pureharm.anomaly`:
 - add `AnomalyLike` super-type that is inherited by `Anomaly`, `Catastrophe`, `Anomalies`, this allows
   for easy non-anomaly catch-all that allows rethrowing.